Create init_vss_gui.ps1
This is a WIP draft of what needs to be done. 
There is more study for vssadmin and diskshadow tools along with powershell and GUI. 
There is also, why use powershell, will it not become a bottleneck for UI ?

I think, this is in requirement gathering and analysis phase. 
I will need more time to design the wrapper and its technology stack.